moved [ to better location
previously the [ was in a different location that was not optimal. In order to maximize the efficacy of our efforts. I took initiative and decided enough was enough. The [ simply had to be moved. It was a hefty task. But I did it anyways. because. that is the kind of person i am. it aligns with my core values of nu-clear-ing. <3